Introduce Interface::Function::Params
